Revert "fix(js): use pure JS SHA256 library, refactor (#471)"
This reverts commit 7b84904d15c307096c717a1c0275ee4dde3a23b0.
