+++ package.json
@@ -24,8 +24,5 @@
     "postcss-import": "^16.1.0",
     "postcss-import-url": "^7.2.0",
     "postcss-url": "^10.1.3"
-  },
-  "dependencies": {
-    "@aws-crypto/sha256-js": "^5.2.0"
   }
 }
